requirements.txt
Django
django-crispy-forms
django-cors-headers
djangorestframework
djangorestframework-gis
psycopg2
django-rest-swagger
djangorestframework_simplejwt
drf-yasg
gunicorn
shapely
